+++ src/test/java/db/MovingDaoTest.java
@@ -3,11 +3,8 @@ package db;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import db.connection.DBConnectionUtil;
-import db.dto.BoardDto;
 import db.dto.MovingDto;
 import java.sql.SQLException;
-import model.Board;
-import model.Camp;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,8 +16,6 @@ class MovingDaoTest {
     @BeforeEach
     void beforeEach() {
         movingDao.remove("moving");
-        movingDao.remove("turn");
-        movingDao.remove("board");
     }
 
     @DisplayName("데이터베이스 접속 확인")
@@ -38,26 +33,5 @@ class MovingDaoTest {
         final var id = movingDao.addMoving(moving);
 
         assertThat(movingDao.findByMovementId(id)).isEqualTo(moving);
-    }
-
-    @Test
-    @DisplayName("보드 저장 확인")
-    void addBoard() {
-        //given
-        final var board = BoardDto.from(Board.create());
-
-        //when
-        movingDao.addBoard(board);
-        final BoardDto findBoard = movingDao.findBoard();
-
-        //then
-        assertThat(board).isEqualTo(findBoard);
-    }
-
-    @Test
-    @DisplayName("턴 저장 확인")
-    void addTurn() {
-        movingDao.addTurn(Camp.BLACK);
-
     }
 }
